Remove the "card" feature
It is too big of a feature and too troubling to maintain.

The "new way" is to just use the card as a regular bank account.
And to pay a card is to transfer money to that account.

This commit deprecates the field chargeable_type. Whenever we remove it
we have to rename chargeable_id to account_id and remove the polymorphic
relation.

I am keeping those fields and the table cards just in case someone wants
to rollback this PR.